+++ ui/src/index.ts
@@ -1,2 +1,2 @@
-import './public-path'
+import './publicPath'
 import '@dashboard'
